+++ core-bundle/src/Resources/contao/classes/DataContainer.php
@@ -1178,7 +1178,7 @@ abstract class DataContainer extends \Backend
 		}
 
 		$return = '
-<form action="'.ampersand(\Environment::get('request'), true).'" class="tl_form" method="post">
+<form action="'.ampersand(\Environment::get('request'), true).'" class="tl_form" method="post" aria-label="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['searchAndFilter']).'">
 <div class="tl_formbody">
   <input type="hidden" name="FORM_SUBMIT" value="tl_filters">
   <input type="hidden" name="REQUEST_TOKEN" value="'.REQUEST_TOKEN.'">