+++ core-bundle/src/Resources/contao/controllers/BackendMain.php
@@ -246,6 +246,7 @@ class BackendMain extends \Backend
 		$this->Template->learnMore = sprintf($GLOBALS['TL_LANG']['MSC']['learnMore'], '<a href="https://contao.org" target="_blank" rel="noopener">contao.org</a>');
 		$this->Template->ref = $container->get('request_stack')->getCurrentRequest()->attributes->get('_contao_referer_id');
 		$this->Template->menu = $container->get('contao.menu.backend_menu_renderer')->render($container->get('contao.menu.backend_menu_builder')->create());
+		$this->Template->headerNavigation = \StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['headerNavigation']);
 
 		$strSystemMessages = \Backend::getSystemMessages();
 		$this->Template->systemMessagesCount = substr_count($strSystemMessages, 'class="tl_');